% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bmi}
\alias{bmi}
\title{Title}
\usage{
bmi(weight, height, name)
}
\arguments{
\item{weight}{number}

\item{height}{number}

\item{name}{name}
}
\value{
bmi
}
\description{
Title
}
\examples{
library(bmi)
bmi(70,173,'John')
}
