% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI.R
\name{BMI_level}
\alias{BMI_level}
\title{Generating body condition level according to gender, age and BMI}
\usage{
BMI_level(data, BMI, gender, age)
}
\arguments{
\item{data}{data.frame}

\item{BMI}{numeric}

\item{gender}{numeric}

\item{age}{numeric}
}
\value{
a data.frame
}
\description{
Generating body condition level according to gender, age and BMI
}
\examples{
data <- data.frame(BMI=18,gender=1,age=9)
BMI_level(data,BMI,gender,age)
}
